networking: Refactor multimess module.
- rename to multihash module and fix usages
