+++ crates/subspace-networking/src/utils.rs
@@ -1,12 +1,11 @@
+pub mod multihash;
 #[cfg(test)]
 mod tests;
 
 use libp2p::multiaddr::Protocol;
-use libp2p::multihash::{Code, Multihash};
 use libp2p::{Multiaddr, PeerId};
 use std::marker::PhantomData;
 use std::num::NonZeroUsize;
-use subspace_core_primitives::PieceIndexHash;
 use tracing::warn;
 
 /// This test is successful only for global IP addresses and DNS names.
@@ -94,14 +93,4 @@ pub(crate) fn convert_multiaddresses(addresses: Vec<Multiaddr>) -> Vec<PeerAddre
             }
         })
         .collect()
-}
-
-pub trait ToMultihash {
-    fn to_multihash(&self) -> Multihash;
-}
-
-impl ToMultihash for PieceIndexHash {
-    fn to_multihash(&self) -> Multihash {
-        libp2p::multihash::MultihashDigest::digest(&Code::Identity, self.as_ref())
-    }
 }
